requirements.txt
azure-functions
mcp[cli]
httpx